requirements.txt
# Telegram Bot requirements
python-telegram-bot==22.5
httpx>=0.27

# Data processing
pandas>=2.0.0
openpyxl>=3.1.0
python-dateutil>=2.8.2

# Google Sheets
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.1.0

# Document generation
python-docx>=1.2.0

# Utilities
python-dotenv>=1.0.0
